+++ js/util.js
@@ -1,0 +1,34 @@
+function getRandomInteger (min, max) {
+  const lower = Math.ceil(Math.min(Math.abs(min), Math.abs(max)));
+  const upper = Math.floor(Math.max(Math.abs(min), Math.abs(max)));
+  const result = Math.random() * (upper - lower + 1) + lower;
+
+  return Math.floor(result);
+}
+
+function createIdGenerator () {
+  let lastGeneratedId = 0;
+
+  return function () {
+    lastGeneratedId += 1;
+    return lastGeneratedId;
+  };
+}
+
+
+function createCommentId () {
+  const previousValues = [];
+
+  return function () {
+    let currentValue = Math.floor(Math.random() * 1000);
+    while (previousValues.includes(currentValue)) {
+      currentValue = Math.floor(Math.random() * 1000);
+    }
+    previousValues.push(currentValue);
+    return currentValue;
+  };
+}
+
+const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
+
+export {getRandomInteger, createIdGenerator, getRandomArrayElement, createCommentId};
